# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orbonomy",
    "version": "1.0.0",
    "description": "Unified AI-powered API platform with 50+ endpoints across accommodation, AI inference, content generation, fact-checking, and real-time data services.",
    "x-guidance": "POST to any endpoint with your data. The server returns 402 with payment-required info. Pay with USDC on Base chain and retry with the payment signature.",
    "contact": {
      "email": "orbonomy@proton.me"
    },
    "x-logo": {
      "url": "https://orbonomy.vercel.app/favicon.png"
    }
  },
  "servers": [
    {
      "url": "https://orbonomy.vercel.app"
    }
  ],
  "paths": {
...
    "/api/llm/translate": {
      "post": {
        "operationId": "llm_translate",
        "summary": "Translation across 50+ languages",
        "tags": [
          "AI Inference"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "target": {
                    "type": "string",
                    "minLength": 1
                  },
                  "source": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "text",
                  "target",
                  "source"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/llm/analyze": {
      "post": {
        "operationId": "llm_analyze",
        "summary": "Multimodal analysis (text, image, code)",
        "tags": [
          "AI Inference"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.100000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "prompt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/llm/rewrite": {
      "post": {
        "operationId": "llm_rewrite",
        "summary": "Rewrite text in different styles and tones",
        "tags": [
          "AI Inference"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "style": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tone": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "text",
                  "style",
                  "tone"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/api/llm/classify": {
      "post": {
        "operationId": "llm_classify",
        "summary": "Text classification: sentiment, intent, toxicity, language",
        "tags": [
          "AI Inference"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.030000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "task": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "text",
                  "task"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/llm/batch": {
      "post": {
        "operationId": "llm_batch",
        "summary": "Batch processing (up to 10 items)",
        "tags": [
          "AI Inference"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.100000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array"
                  },
                  "system_prompt": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "items",
                  "system_prompt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
    "/api/factcheck/summary": {
      "post": {
        "operationId": "factcheck_summary",
        "summary": "Get Wikipedia article summary for a topic",
        "tags": [
          "Fact Check"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic": {
                    "type": "string",
                    "minLength": 1
                  },
                  "lang": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "topic",
                  "lang"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
  }
}